# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sistem Akademik Mahasiswa API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /nilai:
    get:
      operationId: listNilai
      tags: [nilai]
      summary: List grades, optionally filtered by student
      parameters:
        - in: query
          name: mahasiswa_id
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of grades
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NilaiListResponse"
    post:
      operationId: createNilai
      tags: [nilai]
      summary: Add a grade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NilaiInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NilaiDetail"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /nilai/{id}:
    put:
      operationId: updateNilai
      tags: [nilai]
      summary: Update a grade
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NilaiUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NilaiDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteNilai
      tags: [nilai]
      summary: Delete a grade
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Mahasiswa:
      type: object
      properties:
        id:
          type: integer
        nim:
          type: string
        nama:
          type: string
        jurusan:
          type: string
        angkatan:
          type: integer
        email:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - nim
        - nama
        - jurusan
        - angkatan
        - createdAt
        - updatedAt
...
    MataKuliah:
      type: object
      properties:
        id:
          type: integer
        kodeMk:
          type: string
        namaMk:
          type: string
        sks:
          type: integer
        dosenPengampu:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - kodeMk
        - namaMk
        - sks
        - createdAt
...
    NilaiDetail:
      type: object
      properties:
        id:
          type: integer
        mahasiswaId:
          type: integer
        mataKuliahId:
          type: integer
        nilaiHuruf:
          type: string
        semester:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        mahasiswa:
          $ref: "#/components/schemas/Mahasiswa"
        mataKuliah:
          $ref: "#/components/schemas/MataKuliah"
      required:
        - id
        - mahasiswaId
        - mataKuliahId
        - nilaiHuruf
        - semester
        - createdAt
        - updatedAt
        - mahasiswa
        - mataKuliah
    NilaiInput:
      type: object
      properties:
        mahasiswaId:
          type: integer
        mataKuliahId:
          type: integer
        nilaiHuruf:
          type: string
          minLength: 1
        semester:
          type: string
          minLength: 1
      required:
        - mahasiswaId
        - mataKuliahId
        - nilaiHuruf
        - semester
    NilaiUpdate:
      type: object
      properties:
        nilaiHuruf:
          type: string
          minLength: 1
        semester:
          type: string
          minLength: 1
    NilaiListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/NilaiDetail"
        total:
          type: integer
      required:
        - data
        - total
...